Only clean up if live symlink exists

+++ service/build.sh
@@ -22,16 +22,19 @@ fi
 mkdir -p "$DEST/build" "$DEST/deploy"
 cd "$DEST"
 
-echo "Cleaning up $DEST/build"
 cd build/
-old_live="$(readlink ../deploy/live)"
-echo "Currently live: $old_live"
-for file in deploy-*; do
-  if [ "../build/$file" != "$old_live" ] && [ "../build/$file" != "$old_live.log" ]; then
-    echo "Deleting old build $file"
-    rm -rf "$file"
-  fi
-done
+if [ -L ../deploy/live ]
+then
+  echo "Cleaning up $DEST/build"
+  old_live="$(readlink ../deploy/live)"
+  echo "Currently live: $old_live"
+  for file in deploy-*; do
+    if [ "../build/$file" != "$old_live" ] && [ "../build/$file" != "$old_live.log" ]; then
+      echo "Deleting old build $file"
+      rm -rf "$file"
+    fi
+  done
+fi
 
 workdir="deploy-$(date --iso=seconds)"
 logfile="$workdir.log"
